loaders: commonize file open logic
consolidated file-opening into a general-purpose function
for all loaders. Removed duplicate code to improve
maintainability and readability.

issue: https://github.com/thorvg/thorvg/issues/3901

+++ src/loaders/external_webp/tvgWebpLoader.h
@@ -42,7 +42,7 @@ private:
     void run(unsigned tid) override;
 
     unsigned char* data = nullptr;
-    unsigned long size = 0;
+    uint32_t size = 0;
     bool freeData = false;
 };
